chore(release): v1.3.0
- change: split index.js into `lib/` files
- style(esnext): general updates
- test: switch test runner from mocha to node --test
- doc(README): practical makeover
  - badge URL update
