+++ .github/workflows/ci.yml
@@ -3,6 +3,8 @@ name: CI
 on:
   pull_request:
     types: [opened, synchronize, reopened]
+  push:
+    branches: master
 
 env:
   CI: true
@@ -11,14 +13,20 @@ jobs:
   lint:
     uses: haraka/.github/.github/workflows/lint.yml@master
 
+  prettier:
+    uses: haraka/.github/.github/workflows/prettier.yml@master
+    with:
+      branch: ${{ github.head_ref }}
+
   coverage:
     uses: haraka/.github/.github/workflows/coverage.yml@master
     secrets: inherit
 
   ubuntu:
-    needs: [lint]
     uses: haraka/.github/.github/workflows/ubuntu.yml@master
 
   windows:
-    needs: [lint]
     uses: haraka/.github/.github/workflows/windows.yml@master
+
+  macos:
+    uses: haraka/.github/.github/workflows/macos.yml@master
