+++ lib/logger.js
@@ -1,0 +1,14 @@
+'use strict'
+
+let logger
+try {
+  logger = require('../logger')
+} catch {
+  logger = {
+    lognotice: console.log,
+    logerror: console.error,
+    logwarn: console.error,
+  }
+}
+
+module.exports = logger
