+++ package.json
@@ -1,11 +1,12 @@
 {
   "name": "haraka-email-message",
-  "version": "1.2.7",
+  "version": "1.3.0",
   "description": "Haraka email message",
   "main": "index.js",
   "files": [
     "CHANGELOG.md",
-    "config"
+    "config",
+    "lib"
   ],
   "scripts": {
     "format": "npm run prettier:fix && npm run lint:fix",
@@ -13,7 +14,7 @@
     "lint:fix": "npx eslint *.js test --fix",
     "prettier": "npx prettier . --check",
     "prettier:fix": "npx prettier . --write --log-level=warn",
-    "test": "npx mocha",
+    "test": "node --test test/*.js test/lib/*.js",
     "versions": "npx npm-dep-mgr check",
     "versions:fix": "npx npm-dep-mgr update",
     "test:coverage": "npx c8 --reporter=text --reporter=text-summary npm test"
